requirements.txt
requests
python-dotenv
notion-client
markdown
weasyprint